# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gbolix API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        userType:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        acquisitionSource:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
        role:
          type: string
          enum: [owner, admin, freelancer, client]
        isActive:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - name
        - email
        - onboardingCompleted
        - role
        - isActive
        - createdAt
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        phone:
          type: string
        website:
          type: string
        country:
          type: string
        city:
          type: string
        timezone:
          type: string
        language:
          type: string
        avatarUrl:
          type: string
    OnboardingInput:
      type: object
      properties:
        userType:
          type: string
        location:
          type: string
        companySize:
          type: string
        acquisitionSource:
          type: string
      required:
        - userType
        - location
        - companySize
        - acquisitionSource
...